requirements.txt
# AI_MICROSCOPE core dependencies
# Tested on Python 3.11+ - Compatible versions for stable operation

# Deep learning framework
# Relaxed to support currently available TensorFlow 2.x wheels
tensorflow>=2.16.0,<3.0.0

# Computer vision and imaging
opencv-python>=4.7.0,<5.0.0
Pillow>=9.5.0,<11.0.0

# GUI framework
customtkinter>=5.2.2,<6.0.0

# Scientific computing
numpy>=1.25.0,<2.0.0
matplotlib>=3.8.0,<4.0.0

# PDF generation for clinical reports
reportlab>=3.6.0,<5.0.0

# Development and testing (optional)
pytest>=7.4.0,<8.0.0
scikit-learn>=1.4.0,<2.0.0

# Note: The following are built-in Python 3.11 modules (no installation needed):
# - sqlite3 (database)
# - json (configuration)
# - logging (debugging)
# - datetime (timestamps)
# - pathlib (file operations)
# - threading (concurrency)
# - tkinter (GUI - customtkinter dependency)
# - csv (data export)
# - io (file handling)
# - typing (type hints)

# Optional packages for enhanced functionality (uncomment as needed):
# tensorflow-lite-runtime>=2.13.0  # For edge deployment
# onnx>=1.14.0  # For model format conversion
# scikit-image>=0.20.0  # For advanced image filters
# scipy>=1.10.0  # For scientific functions
# psutil>=5.9.0  # For system monitoring
# memory-profiler>=0.60.0  # For memory analysis
# black>=23.0.0  # For code formatting
# flake8>=6.0.0  # For code linting
